+++ R/package-SHT.R
@@ -11,8 +11,8 @@
 #' @noRd
 #' @aliases SHT-package
 #' @import Rdpack
+#' @importFrom flare sugm
 #' @importFrom pracma pinv strcmp
-#' @importFrom fastclime fastclime
 #' @importFrom utils packageVersion
 #' @importFrom stats pt sd cov pchisq pf median aov cov2cor pnorm var runif shapiro.test rgamma rlnorm cor ppoints qnorm dist rnorm qnorm punif uniroot
 #' @importFrom Rcpp evalCpp